+++ local_app/TweetCatAppMac/TweetCatApp/TweetCatApp.swift
@@ -10,31 +10,6 @@ class AppDelegate: NSObject, NSApplicationDelegate {
         }
 }
 
-// 放到任意初始化早、且只注册一次的地方（比如 ContentView.onAppear 或 App.init）
-final class NativeMessageReceiver: ObservableObject {
-        static let shared = NativeMessageReceiver()
-        @Published var lastPayload: [String: Any] = [:]
-
-        private init() {
-                DistributedNotificationCenter.default().addObserver(
-                        forName: Notification.Name(
-                                "com.tweetcat.nativeMessage.incoming"
-                        ),
-                        object: nil,
-                        queue: .main
-                ) { note in
-                        if let payload = note.userInfo?["payload"]
-                                as? [String: Any]
-                        {
-                                print("收到扩展消息:", payload)
-                                self.lastPayload = payload
-                        } else {
-                                print("收到扩展消息但无 payload")
-                        }
-                }
-        }
-}
-
 @main
 struct TweetCatApp: App {
         @NSApplicationDelegateAdaptor(AppDelegate.self) var appDelegate
